Update patch set 1
Patch Set 1:

I pulled down the .aar files for RecyclerView in both the support 28.0.0 release and the AndroidX 1.0.0 release, and it appears that the getBackgroundThreadExecutor method is annotated with nonnull on both, so the check that it has become nullable is actually valid and means that 1.0.0 -> 1.1.0 is actually a breaking change.  :(

Not sure what we should do about that.  I'll look at the code to see if it really seems like it needs to be @nullable, and will consult with ccraik@ (the author who made it so) for his opinion.

Or we could just ship 1.1.0 beta with a breaking change in it... but that seems to go against protocol.

Patch-set: 1
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>