Update patch set 1
Patch Set 1:

I think this might be too small a chunk for the review to be super useful: when this was implemented in stages like this before the earlier parts ended up being revised a lot once the later parts were implemented. It seems like the complete implementation should be small enough to review in one piece?

Patch-set: 1
Reviewer: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>