Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> LGTM.
> 
> Unrelated - but why do we assume all the online big cores are contiguous with cpu0?
> 
> while [ -f ${CPU_BASE}/cpu${cpu}/online ]; do cpu =$(($cpu + 1))
> 
> vs
> 
> for cpu in $(find "$CPU_BASE" -name cpu* | grep "cpu[0-9]+") ; then

That line isn't iterating over all online big cores, it's iterating over every core. It looks for each place it finds the 'online' file (which may be big or small, and online or not based on the contents of the 'online' file).

Patch-set: 1
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>