Update patch set 3
Patch Set 3:

(2 comments)

Looks good. Would be interesting to see if this increases performance. You don't need to do it in this CL, but it would be a good idea to instrument this code with trace markers so we can track performance regressions.

Patch-set: 3
Reviewer: Gerrit User 1377888 <1377888@85c56323-6fa9-3386-8a01-6480fb634889>