Stability fixes
1. On some legacy HAL devices, the image reader from a previous capture
session can persist after a device is closed. This can cause a crash
when a new preview stream is started.

This patch creates a temporary dummy surface when a preview stream is
closed on these devices to prevent this.

2. Remove unnecessary closeAllCameras calls before each test run.
This is unnecessary and has the side effect of sometimes causing
extra close events that can stall Espresso tests.

This patch removes these unnecessary calls.

3. Add a variable for preview buffer for Espresso tests and increase the
value to 1500ms. On some devices the Camera 1 API can be overwhelmed if
a chain of requests hits it to quickly.

Test: gradlew camera:integration-tests:camera-testapp-timing:connectedCheck --info --daemon
Test: manually tests on Pixel sailfish 8.0.0
Change-Id: I394b8b3ee474bf16a543ed43b033f854dd323620