Workaround for ImageReader persistence bug
On some legacy HAL devices, the image reader from a previous capture
session can persist after a device is closed. This can cause a crash
when a new preview stream is started.

This patch creates a temporary dummy surface when a preview stream is
closed on these devices to prevent this.

Test: gradlew camera:integration-tests:camera-testapp-timing:connectedCheck --info --daemon
Test: manually check the switch tests on Pixel sailfish 8.0.0
Change-Id: I394b8b3ee474bf16a543ed43b033f854dd323620