Make selection controls API be foo/onFooChange
We tried to make value / onClick API for selection controls, but it
looks not really right in terms of both declarative and usability.

The standard practice in all the reactive frameworks (flutter / react)
os to have foo: FooType / onFooChanged : (FooType) -> Unit methods for
components that involve input.

Also, this API makes +state deconstruction usage so much better and
shorter, e.g.
val (state, onStateChange) = +state { true }
Checkbox(state, onStateChange, customColor)

Part of this change is making Checkbox respect this structure as well as
be some simple to use with Boolean overload instead of ToggleableState

This CL does this for Checkbox and Switch.

Bug: 133457277
Bug: 133456247

Test: ./gradlew ui-material:connectedCheck
Change-Id: I7cf7bc6c5f6e39bea09d28ad81b2f68b8c3e906a