Fix incorrect pop animation in multiple transactions
When a fragment is added by a pop and then removed in back to back
transactions, we should use a normal exit animation instead of a popExit
animation.

This change swaps the normal and pop animations of all of the Ops in a
transaction, if the pop transaction is not the last one being executed.

Test: all tests pass, and tested in app
Bug: 111659726
Change-Id: I18eefc2fdb4fc4c27c467212cc2e68beb20ea00b