Update patch set 4
Patch Set 4:

I added new methods to the WebViewAssetLoader API to host app's internal storage. It's disabled by default unlike assets and resources.

I haven't added any unit or integration tests yet till we agree if this is the final shape we want for the API.


> Patch Set 1:
> 
> I think this might be too small a chunk for the review to be super useful: when this was implemented in stages like this before the earlier parts ended up being revised a lot once the later parts were implemented. It seems like the complete implementation should be small enough to review in one piece?

Well, it is a small independent utility method that does not affect the API. But Anyway, I added the change to the main API here as well instead of a separate CL.

Patch-set: 4
CC: Gerrit User 1101490 <1101490@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1402701 <1402701@85c56323-6fa9-3386-8a01-6480fb634889>