Dynamic target animation
This CL introduces the concept of AnimatedValue. Its main responsibility is
to hold a value (of type T), and whever that value changes, it animates the
change.

AnimatedFloat is a AnimatedValue that holds a float type value. It tracks
not only the value but also the velocity of the animation when value changes.
AnimatedFloat is perfect for handling fling gestures.

Test: included tests in the CL & 4 demos for visual inspection

Change-Id: I83ae335302829ce6253259647b9f2b65bc8a0f65