Create API for using Room with a pre-populated database.
The builder method createFromFile() allows the user to configure Room
to use a pre-packaged database if the DB is being opened for the
first time ever. By default the DB file has to be placed in the
'assets/' folder of the APK, however if fromExternalAsset is set then
Room will try to locate the file in external storage. As long as the
right permissions are available Room will be able to utilize the
external storage path.

If the pre-packaged database file is an older version, then Room will
go through the upgrade flow as usual.

The pre-packaged database does not need to contain Room metadata table,
upon being opened Room will validate the schema and create it.

Bug: 62185732
Test: room:integration-tests:room-testapp:cC
Change-Id: I5d84a48f8511bcf3adaaaf6b8bf4073843f01ac7