Create APIs for using Room with a pre-populated database.
The builder methods createFromAsset() and createFromFile() allows the
user to configure Room to use a pre-packaged database if the DB is
being opened for the first time ever. The fromAsset version requires the
db file to be placed in the 'assets/database/' folder of the APK,
while the fromFile allows the DB file to be located in external storage.
As long as the right permissions are available Room will be able to
utilize the external storage path.

Bug: 62185732
Test: room:integration-tests:room-testapp:cC
Change-Id: I5d84a48f8511bcf3adaaaf6b8bf4073843f01ac7