Use local HashSet instead of field BitSet for tables invalid statuses.
State touched within the Refresh Runnable has to be thread safe and
BitSet is not thread-safe and was not synchronized externally. It is
best to take some performance penalty by not having a field. Instead
we create a HashSet of invalid tables locally for every refresh. This
keeps things simple and we avoid the complexity of synchronizing the
invalid statuses.

Bug: 133457594
Test: Write a multi-threaded test? Lets just say all tests pass. :)
Change-Id: Iddedbf17cef9daf264569b027c8ff199c9813267