Remove T?.orFromTheme() theming methods in favour of non-null params with defaults
Previously we used syntax like
Button(color : Color? = null) {
   val finalColor = color.orFromTheme { primary }
}
But that makes it harder to follow what value will be used by default and hides the dependency on the theming ambient.
We can make it more explicit with this syntax:
Button(color : Color = +themeColor { primary })

Even that approach forces us to duplicate defaults if we just wrap our Button and want to proxy the params
FancyButton(color : Color = +themeColor { primary }) { Button(color) }
It still seems beneficial because
a) FancyButton can indeed have different defaults
b) It is even harder to find what is the default value for FancyButton when it is nullable as to figure it out you have to dig into the Button internals

Bug: 134046513
Test: tests are passing, demo app is working
Change-Id: I9e548132cdab4ed8b9ee5ef4d2a9669d5f245c27