Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> Quick question, do we still need to suppress these tests since we have: https://android-review.googlesource.com/c/platform/frameworks/support/+/970303 ready for merge?

I think we want to add these as well because the presubmit framework is not really set up to handle Antelope hammering the cameras. If a device gets in a bad state, it can cause all the other future camera tests to fail.

I think we need to add @Suppress to all the tests. Then merge 970303 and the other outstanding ones.

Patch-set: 1
Reviewer: Gerrit User 1269804 <1269804@85c56323-6fa9-3386-8a01-6480fb634889>