Support many-to-many with @Relation using @JoinEntity
Expand @Relation to contain a new method, 'joinEntity' of type
@JoinEntity that can be used to specify a join table for resolving
relationships, mainly many-to-many.

Upon using @JoinEntity the query
generated by the RelationCollector is of the form of:
SELECT *, <join-parent-column>
  FROM <join-table>
  JOIN <entity-table> ON (<join-entity-column> = <entity-column>)
  WHERE <join-parent-column> IN (:args)

@JoinEntity supports both an @Entity or @DatabaseView, the join table
itself doesn't have to strictly model a many-to-many relation. Both
parentColumn and entityColumn are optional and will be assumed to be
the same as the ones defined in @Relation, otherwise those methods can
be used to specify arbitrary columns in the jump table.

Bug: 69201917
Test: ManyToManyRelationTest & PojoProcessorTest
Change-Id: I967cf77252391379377ebe0ebfeb23a4fd881ca5