Support many-to-many with @Relation using @Junction
Expand @Relation to contain a new method, 'associateBy' of type
@Junction that can be used to specify a junction table for resolving
relationships, mainly many-to-many.

Upon using @Junction the query
generated by the RelationCollector is of the form of:
SELECT *, <junction-parent-column>
  FROM <junction-table> AS _junction
  JOIN <entity-table> ON (<junction-entity-column> = <entity-column>)
  WHERE <junction-parent-column> IN (:args)

@Junction supports both an @Entity or @DatabaseView, the Junction table
itself doesn't have to strictly model a many-to-many relation. Both
parentColumn and entityColumn are optional and will be assumed to be
the same as the ones defined in @Relation, otherwise those methods can
be used to specify arbitrary columns in the junction table.

Room will also give a warning for those columns in the junction table
that are not part of a foreign key.

Bug: 63736353, 69201917
Test: ManyToManyRelationTest & PojoProcessorTest
Change-Id: I967cf77252391379377ebe0ebfeb23a4fd881ca5