Ensure ActionBar is not updated when destination is floating window
When showing a dialog destination, the actionbar updates to show an up
arrow. This behavior is unexpected and it is not possible to use the
back button. The action bar should remain the same when interacting with
any floating window.

This change adds isFloatingWindow field to NavDestinations to allow them
to identify whether they are a floating window.
ActionBarOonDestinationChangeListener checks the floating window status
to determine if the up arrow should be shown.

Test: Tested in integration app. ./gradlew checkApi
BUG: 133600763
Change-Id: Ie4efab6c265eca13583bd21615f8d571c2236443