Drop data in MetadataImageReader when data becomes stale
Modified FakeImageReaderProxy and FakeImageProxy to handle closing
behavior and stalling behavior on max images being pushed.

Added data dropping in MetadataImageReader when the number of ImageProxy
reaches maximum acquired so that there will always be room for 1 more to
be acquired. Also drop the ImageProxy or ImageInfo with smaller
timestamp than the smallest timestamp of the other data to be matched.
This is because we assume the data coming in, whether ImageProxy or
ImageInfo, is to be monotonically increasing in timestamp values.

In addition acquire all images from ImageReaderProxy when images are
available because there might be multiple waiting to be acquired.

Bug: b/130193125
Test: ./gradlew camera:camera-core:build
Change-Id: I3eba9e95aff24f51dbcbc52c1631704b7135c6f7