Add SingleChildLayout
The CL adds a SingleChildLayout widget, built on top of Layout, which
expects one Measurable child. When multiple Measurables are discovered
during composition, these are wrapped in one combined Measurable, which
has a Stack-similar behavior: it sizes itself to the largest child
Measurable and placing its corresponding Placeable will place each
child at the top left of the layout.

Bug: 134388814
Test: ran all Layout tests
Change-Id: I007a02738d2753f1042f8ad287c8e878c49dd728