Fix unit test fail - Camera2ImplCameraXTest.bind_unbind_loopWithOutAnalyzer() / bind_unbind_loopWithAnalyzer()
Root cause:
The issue is caused by the timing of the onOpened() callback.
Under the test pass condition, the unit test does not actually wait
for the second onOpened(), instead it passes the second verification
directly because the mock callback already has an Opened() by the
first run. So it passed the test by coincidence.
If it waits for the second camera open/close, it will fail as shown
by this bug.

Solution:
Do not use the same mock callback, create new mock callback for each run.

Bug: 134360074
Test: ./gradlew camera:camera-camera2:connectedAndroidTest -Pandroid.testInstrumentationRunnerArguments.class=androidx.camera.camera2.impl.Camera2ImplCameraXTest#bind_unbind_loopWithOutAnalyzer ; ./gradlew camera:camera-camera2:connectedAndroidTest -Pandroid.testInstrumentationRunnerArguments.class=androidx.camera.camera2.impl.Camera2ImplCameraXTest#bind_unbind_loopWithAnalyzer
Change-Id: I4eff4ab757c3b6f8d9e83527ed71e6022321efc1