Update patch set 37
Patch Set 37:

(8 comments)

In general, be careful that we only change kotlin tests in the conversion if we're making things better. Changing `.size` -> `.getSize()` is a source breaking change for the worse, not sure about `V : Any. metalava should help us prevent this, would be good to put together notes on improvements that could be made here, and we can meet with the metalava owners to talk about it.

Patch-set: 37