Allow using POJOs in @Insert, @Update and @Delete method parameters.
Add a new annotation to Room, called @AsEntity that allows annotation
method parameters in DAO methods with @Insert, @Update and @Delete.
Room will interpret that you want the use the Pojo as a partial entity
for performing the Dao method.

If the method is an INSERT, we allow any Pojo with a subset of columns
that will not cause a NOT NULL constraint. If the real entity has a
auto generated primary key, then that column is optional too.

If the method is an UPDATE or DELETE then we allow any Pojo with a
subset of columns regardless of the schema default values. The only
requirement here is that the Pojo must have fields matching the
primary keys of the entity.

Bug: 127549506
Test: ./gradlew room:room-compiler:test \
                room:integration-tests:room-testapp:cC
Change-Id: Ie0477d91bb92f799486ff672b20d4564769903d2