Update patch set 7
Patch Set 7:

Is there a way to get this working with a negative margin? I am trying it but with a negative margin is only working correctly after a little scroll. Is there a way to apply the translationX/Y to the previous and next items without this scroll?

Patch-set: 7
CC: Gerrit User 1510105 <1510105@85c56323-6fa9-3386-8a01-6480fb634889>