Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Is this just a simple upstream merge?  Overall, seems about accurate to me

In Gerrit to the right of the "Files" button and to the left of the "Patchset 4" dropdown is another dropdown that can change the base that the diff is shown against. One possible value is "Auto Merge" which shows the difference between this change and `git merge`.

Patch-set: 4
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>