Update patch set 7
Patch Set 7:

> Change DatabaseVerifier so it can verify CREATEs as well.
The verifier already executes all the CREATE TABLE to later verify the queries in the Dao, I say just make sure that if an error shows in the verifier while creating the tables then be sure to print a nice message with the error and with the @Entity element that is related.

For the inconsistency between @ColumnInfo(defaultValue = "...") and the field's value at initialization, I think that maybe all we need is clear documentation that the field's initialized value does not affect the actual schema's default value, they are kinda two different things.

Schemas default value affects @Query("INSERT...") statements and eventually Partial Entities in @Insert, @Update and @Delete methods (when the column is missing in the partial entity). It also helps with migrations, even though also a change in default value requires a migration.

Meanwhile field init values (or Kotlin default values) affect when you query and parse the Pojo and that field's column is not in the result. It happens to affect @Insert, @Update and @Delete too because that is the value read by the Dao during the operation, but it can be anything after the object is initialized.

I guess if there is something to recommend is to not have both, since that is indeed confusing, but both have their differences that might or might not make them useful in different cases.

Patch-set: 7
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0