Ensure FragmentNavigator pops Fragments properly
If a navigate and pop are executed on a FragmentNavigator in the same
FragmentTransaction and the FragmentManager back stack is empty, the pop
never occurs, leaving the wrong destination on top of the back stack.
This results in the incorrect Fragment being shown at the end of the
transaction.

This change removes the check for the FragmentManager back stack being
empty and enqueues the pop anyway. The FragmentManager handles popping
with an empty backstack.

Test: Adjusted test, old test pass
BUG: 133832218
Change-Id: I3d33df940dc58b548c44f38b5433a4e938621527