Update patch set 1
Patch Set 1:

Thinking about this some more, I think this needs a larger change since the attached bug also talks about order insensitivity for query parameters. This seems to imply that the initial pattern match should only take into account the Uri without query parameters (essentially cutting it off at the '?' if it exists) and the mapping of query parameter names to argument names kept for later. That mapping would then be used to fill in the arguments and do the check to ensure that all arguments named in the mapping that don't have default values are filled in from the query parameters.

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>