Update patch set 3
Patch Set 3:

I like much of the solution in this, but I'm not sure about the SingleChildLayoutNode because it makes it difficult to mix single child nodes with multiple child nodes and still save on LayoutNodes.

That said, I really need to do a full performance evaluation to see if it is worth saving anything as opposed to making the tree deeper by adding full Stack() elements into the layout. My gut says we should do something like this, but my brain tells me to test and verify. 

Can we save this CL until after those tests are done?

Patch-set: 3
Reviewer: Gerrit User 1060493 <1060493@85c56323-6fa9-3386-8a01-6480fb634889>