Update patch set 1
Patch Set 1:

Just checking, instead of "instrumentation", what does everyone think of "inspector" or "inspection" instead? Since that seems to be the vocabulary we've been leaning towards more and more with this feature?

"instrumentation" sounds pretty general to me, focusing more on the fact that we have some input bytes that we modify, vs. "inspector", which focuses more on the fact that we plan to inspect / modify the state of an app related to some library. Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1047533 <1047533@85c56323-6fa9-3386-8a01-6480fb634889>