Update patch set 1
Patch Set 1: Autosubmit+1 Presubmit-Ready+1

We've been lucky with this bug because room-runtime brings in androidx.collection so it was never the case in the wild where ArrayMap was not found in the classpath and keyAt() and valueAt() always worked.

But that is not the same in goog3 where all dependencies are strictly declared.

Patch-set: 1
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Presubmit-Ready=+1