Use AndroidViewModelFactory as the default Factory
Make ComponentActivity and Fragment override
getDefaultViewModelProviderFactory() to properly
provide an AndroidViewModelFactory.

Updates the 'by viewModels()' Kotlin extension
to use the same factory for the default.

For 'by activityViewModels', we specifically use
the activity's default factory, rather than the
Fragment's default factory to ensure that the
behavior matches calling 'by viewModels' in the
activity.

Test: updated tests pass
BUG: 110936593
Change-Id: Ie5871e23a6360bdaab4979789be40005258e686c