Update patch set 3
Patch Set 3:

A concern with the ordering of priority.

If the SessionConfig has higher priority than CameraEventCallbacks, then it means that the OEM might not be able to override certain parameters that we set in the UseCase. So wouldn't it make sense for CameraEventCallbacks to have higher priority?
Or is there something that I'm missing?

Patch-set: 3
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>