Update patch set 1
Patch Set 1:

idk if you moved all but we should leave some tests behind or write new ones for the deprecated stuff. We need to ensure that we are not breaking it, there is a good chance that people ship libraries that already use them.

Patch-set: 1
Reviewer: Gerrit User 1060337 <1060337@85c56323-6fa9-3386-8a01-6480fb634889>