Create custom ViewGroup for Fragments
Existing ViewGroups can turn on view animations (and transitions) by
setting animateLayoutChanges to true. When attempting to use both view
and fragment animations, it is possible to observe unexpected behavior
because the animation systems are not synced very well. aosp/962786
disables fragment animations if animateLayoutChanges is true on its
view.

This custom ViewGroup disables view animations (and transitions). If a
developer attempts to set transitions on the view, either through
setting animateLayoutChange to true or explicitly using
setLayoutTransition(), an UnsupportedOperationException will be thrown.

Test: Added FragmentLayoutTest
BUG: 116257087
Change-Id: I7498950b502e8c51f941c8c1e59dbfe11649b6bc