Update patch set 1
Patch Set 1:

Was there a particular reason you chose the FragmentLayout name? I'd to avoid developers getting the impression that they can treat this as a general purpose layout but instead just as a container for Fragments - perhaps FragmentContainer is closer to that?

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>