Update patch set 2
Patch Set 2:

(2 comments)

I understand the need for this, but I would advocate for not adding `name` parameters to any component above `Layout`. It seems like we are adding `name` to a LOT of highly used public components here. These parameter lists are getting big as-is, and I don't think we want this type of naming to be propagating upwards too much.

We are already working on some tools to make the composition hierarchy more understandable at dev-time, and I don't want this to take too much of its place in terms of public API surface area.

Patch-set: 2
CC: Gerrit User 1493285 <1493285@85c56323-6fa9-3386-8a01-6480fb634889>