Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 4:
> 
> (1 comment)
> 
> It seems like this always waits for a fixed duration, meaning the test will always run for the same amount of time (2000ms + IDLE_TIMEOUT_MS), and there is no guarantee the test will be finished by then.
> 
> Is there any way you can do this with CountingIdlingResource instead?

Oh, yes, the current implementation seems not good...
There is a update for this.

The reason that I need to wait for a little bit time is because the test need to switch the extension effect after the CatpureSession is created.
In the updated test implementation, I would try to take a picture and wait until the onImageSaved callback was invoked. <= My thought is ... the issueSingleRequest will only start after CaptureSession is opened. To wait until the onImageSaved callback can make sure the CaptureSession is created.


So, the new code will:
(1) Default hide the TakePicture button, and show the TakePicture button after the ImageCaptureUseCase instance ready.
(2) Wait for the TakePicture button visible, and click the take picture button to take a picture.
(3) Wait for the onImageSaved() (by using the countingIdlingResource), and then switch to next extension effect

Patch-set: 6
Reviewer: Gerrit User 1475857 <1475857@85c56323-6fa9-3386-8a01-6480fb634889>