Update patch set 1
Patch Set 1: Code-Review-1

First of all, I don't want to maintain ListenableFuture (LF).
Once we have the copy of LF and release it, when we move back to LF lib without code copy, it will be a breaking change.

Actually, I haven't followed up the issues on LF and VersionedParcelable (VP) in minute detail. Nick, or Alan, can we have a sync-up meeting for the current status of LF (and VP in the mean time)?

If LF and/or VP are not a recommended way to use in public APIs anymore, although I'm not very happy about it, I feel that releasing media APIs without the use of of LF and/or VP is the right way to go (even it requires several more months delay).

Thanks,

Patch-set: 1
Reviewer: Gerrit User 1018371 <1018371@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1