Deprecate Preference#onInitializeAccessibilityNodeInfo
In general, developers should not customize the View's AccessibilityNodeInfo
unless they are customizing the View. If the customization is common to a specific
preference class, the preference class should be modified. If there is really a need to
customize the View's AccessibilityNodeInfo, developers can use the onBindViewHolder method
of the Preference to get a reference to the view associated with the Preference and set an
AccessibilityDelegate for the View.
Bug: 78136923
Test: no specific test needed

Change-Id: I2f6ae3d21aa43b7b2ac2d1d6fa66d58239fadf99