Update patch set 1
Patch Set 1:

Hello there,

So following conversation with Toby, we thought that rewriting the AssetLoader API to make it more generic with the developer having the ability to register custom handlers for different paths.

I haven't updated the java doc or the unit tests for this (so please ignore it for this review iteration). I'll definitely do this after we reach an initial agreement if this is a thing we would like to continue with.

I'm not asking for a in-depth review, I just want to know your thoughts about this new API.

Thanks

Patch-set: 1
Reviewer: Gerrit User 1141809 <1141809@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1101490 <1101490@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1402701 <1402701@85c56323-6fa9-3386-8a01-6480fb634889>