Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Thanks George! Also based on what Leland said above, is this API going to be temporary?

Yeah, I was making a proposal here. I also didn't like the way it was adding "name" as public API to commonly-used components.

There is a new way around this that I can think of -- we can add a "name" node to the tree as a DataNode.

I'm going to abandon this change. I don't think this debug interface is really something we want in the public API.

Patch-set: 2
Reviewer: Gerrit User 1060493 <1060493@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0