Make it so that Draw() can let children not draw.
Feedback from API council was that developers may want
to exclude children from being drawn. This CL allows
the developer to choose not to call drawChildren().

Draw(children) now takes an extra "drawChildren"
argument, making sure that the developer knows that
she must call drawChildren. With drawChildren()
in the receiver scope, it hid the need from the
developer, so I thought it was best to be explicit.

Test: ui-framework:cC ui-platform:cC

Change-Id: I5efc8e29fedb8077348291ebcd7a762fba3f0ac4