Update patch set 1
Patch Set 1:

This is binary compatible but may not be source compatible. For this specific example, you would have to have a raw Class type and be passing it to JobIntentService. What's likely the case is that you're passing a class literal like Foo.class which is going to continue to work just fine.

Patch-set: 1
Reviewer: Gerrit User 1232209 <1232209@85c56323-6fa9-3386-8a01-6480fb634889>