Update patch set 1
Patch Set 1:

This is binary compatible but my not be source compatible. For a source breakage, the person would have to have a static final Class type and be using that as the value. The overwhelmingly common case of using a class literal like Foo.class will continue to be source compatible.

Patch-set: 1
Reviewer: Gerrit User 1232209 <1232209@85c56323-6fa9-3386-8a01-6480fb634889>