Ensure FragmentContainView passes down WindowInsets
In a FrameLayout, when using fitsSystemWindows=true, the layout consumes
the insets and fails to pass them to child views. This can cause
interactive elements to be hidden under SystemWindows.

This change adds an OnApplyWindowInsetsListener that when dispatched,
returns the insets, ensuring they are not consumed by the layout and can
be applied to child views.

Test: Tested visually in app
BUG: 133258137
Change-Id: I8fd58230e8339d417ab2bcb3a89ae2790bcc8172