Implement first version of PagedSource
First step towards migrating to PagedSource interface. This a new
api replacing DataSource in paging 3.0. It combines the functionality of
different subclasses of DataSource under one class and differentiates
available optmizations by its KeyProvider.

As the first step, this change simply switches Pager to use the
PagedSource api over DataSource and a PagedSourceWrapper was introduced
to migrate the existing Listenable*DataSources to the PagedSource api.

Test: ./gradlew paging:paging-common:check
Test: ./gradlew paging:paging-runtime:cC
Change-Id: I1d04abdb857f686d4bde3b12dedae5cfd434dac5