Update patch set 6
Patch Set 6:

(6 comments)

Addressed comments.

For the algorithm selection for both EncryptedFile and EncryptedSharedPreferences, API level is intentionally mentioned to note that it will work differently on older devices. I'm curious what Thai's on this are? I'm not sure if we need to be more explicit for older devices.

Patch-set: 6
Reviewer: Gerrit User 1060006 <1060006@85c56323-6fa9-3386-8a01-6480fb634889>