Update patch set 4
Patch Set 4:

PS4 has an API change for IdentityCredential.getRequest() - we now always require a non-null CBOR request message ... filtering can be done by the app passing in a Collection<RequestNamespace> to specify which items in the request message to _not_ get.

Yes, this is a little backwards and annoying but it has to work this way because the reader can sign the request message and thus the app cannot modify it.

Patch-set: 4
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>