Fix flaky MCV#setPlayer
Since ExoPlayerMediaPlayer2Impl#getDuration should not be called when
current media item is null, this CL checks if the media item in
onCurrentMediaItemChanged() is null and if yes, does not call it.
Instead, it will reset the metadata UI in VideoView/MediaControlView.

Also, since
ExoPlayerMediaPlayer2Impl#getCurrentPosition/getBufferedPosition
should not be called when current player state is IDLE, added code to
check current state in PlayerWrapper.

Bug: 135641665
Test: ./gradlew :media2:media2-widget:check \
                :media2:media2-widget:connectedCheck

Change-Id: I0688a1289d6a91f395f7ce9be116340998b2d16a