Draw exit animations first in FragmentContainerView
ViewGroups always draw exiting view animations last. For Fragments, this
means that the exiting Fragment is always on top of all others. There
can be no interaction with entering fragments, and there is the
possiblity of a "pop" when the exit animation finishes to reveal the
Fragment underneath.

FragmentContainerView overrides the drawChild method to force all
disappearing children to be drawn first. It overrides public methods
from ViewGroup to build the proper list of transitioning views and
disappearing child views. Then a reference view is used to indicate when
the disappearing views should be drawn or skipped.

Test: tested visually in app, ./gradlew checkApi
BUG: 37036000
Change-Id: I054eaf4ae234258bc7abec22c4c8e559a2c6be71