Update patch set 6
Patch Set 6:

Note that I made a modification to the ProcessingSurfaceTexture to use a SingleImageProxyBundle which is in aosp/987425 so that it no longer needs to keep track of a SettableImageProxyBundle which needs to be reset. Instead the SingleImageProxyBundle is recreated each time.

Patch-set: 6
Reviewer: Gerrit User 1482448 <1482448@85c56323-6fa9-3386-8a01-6480fb634889>