Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > I think everything is in radians in other parts of the API, including rotate.
> 
> Is this intentionally different from platform-native for some reason? View & Canvas all appear to be in degrees. Seems like this would be a source of friction to migrate to Compose.

Sorry I wasn't clear. I don't mind that it is in degrees. I mind that only _some_ of the APIs are in degrees.

It is much easier to type and reason about "canvas.rotate(90)" than "canvas.rotate(PI/2)" IMO. But I'm willing to live with either radians or degrees as long as it is consistent.

Patch-set: 1
Reviewer: Gerrit User 1060493 <1060493@85c56323-6fa9-3386-8a01-6480fb634889>