Make fingerprint helper fragment use default error string
Currently, `FingerprintHelperFragment` can pass a null error string to
`AuthenticationCallback.onAuthenticationError` in a couple of ways:
- By forwarding a null error string from the fingerprint manager
authentication callback
- By mapping an unknown error code to null in `getErrorString`

This commit changes the behavior in both of these cases to use the
existing default error message rather than a null string. This should
prevent crashes in Kotlin implementations where `errString` is
enforced as being non-null. It also adds log statements for both of
these cases to help with future debugging of unknown error codes.

Test: build + manual
Fixes: 128350861
Change-Id: Ibd07b51bc810b0d7914969ba830cd9eabad07cbd