Call onVideoSizeChanged when media is prepared
MediaController returns the VideoSize value of the previous media
item if the current media item does not have a video
(e.g. music file). This CL fixes this by making MediaPlayer retrieve
VideoSize when it is prepared and call onVideoSizeChanged.

Also, made PlayerWrapper call onVideoSizeChanged when it is
attached to VideoView.

Bug: 136039839
Test: ./gradlew :media2:integration-tests:testapp:installDebug on API
28, 27 devices
Change-Id: I617fc8e27ab3554dd078b3036548fc9239eedf92