Populate setScrollDeltaX/Y for RecyclerView
getScrollDeltaX/Y are populated by view for TYPE_VIEW_SCROLLED
accessibility events. The events are sent in onScrollChanged,
where the difference in old and the new scroll positions are
parameters and are used to populate the events. RecyclerView
passes only the new scroll position, as both the old and the
new parameters. This should be fixed.

Test: currently none
Bug: 131267612

Change-Id: I105553a1b7c34b11e1227d28ab6be2ccae4c42e6