Animate Toolbar Changes in NavigationUI
When animateLayoutChanges is set on a Toolbar, there are errors in how
the text within the toolbar is displayed if it has an arrow. The text
leaves a big space on destinations with no arrow.

This change adds a boolean to the setNavigationIcon method in
AbstractAppBarOnDestinationChangedListener to indicate whether the icon
should be shown as a drawer indicator. Toolbars then can use a
TransitionManager to animate at the appropriate time.

Test: Tested visually in app
BUG: 131403621
Change-Id: I97ff213bdeb6da3282287d9b572f33c74cf9e9e7