Add benchmarks module.
The module is child of integration-tests module which is mainly meant to
contain correctness tests (such as asserting metrics counts) to ensure
validity of the benchmarks itself.

The idea is that each benchmark would have a TestCase class containing
most of the logic and we would run it twice. Once in benchmarks and
second time in the integration tests. In the integration tests we can do
variaty of asserts including easier debugging.

I'm also trying to introduce the ability to fake the choreographer used
in compose to have more control over all the events happening.

Test: Added
Bug: 127479944
Change-Id: I2e24affb695ef04c5e9814c5db97ccfa8d5a1241