Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3: Code-Review+1
> 
> Q) What's the expected behaviour of skipToNextPlaylistItem? Should it call onCurrentMediaItemChanged? If so, please modify corresponding test cases such as testSkipToNextItems to check the callback is called. (other skipTo* methods as well)

Right. According to the SessionPlayer#skipTo{Previous,Next,*}PlaylistItem, onCurrentMediaItemChanged should be called and it so test cases should check it.

And it would be great if we mention it in the doc of SessionPlayer.PlayerCallback#onCurrentMediaItemChanged as well.

Thanks Jinseok for finding/fixing this!

Patch-set: 3
Reviewer: Gerrit User 1018419 <1018419@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1