Update patch set 3
Patch Set 3:

Hey Siyamed@. Since you are still working on this, there is also a mistake in the doc of TextPainter.didExceedMaxLines.
If [maxLines] is null, this is true if [ellipsis] is true and there was a line that
     * overflowed the `maxWidth` argument passed to [layout]; otherwise it is false.

Where the "ellipsis" is not valid anymore, can you also fix this?

Patch-set: 3
Reviewer: Gerrit User 1362035 <1362035@85c56323-6fa9-3386-8a01-6480fb634889>