Separate Draw with children from Draw without children.
Draw with children requires creating an extra object and
was creating at least two objects on each draw call.
This CL separates the draw with children from the draw
without children. This helps API because `drawChildren()`
is no longer accessible when there are no children and
also eliminates creating an extra object.

This also moves the receiver scope to core so that an
extra level of receiver scope indirection isn't needed.
We also now cache the receiver scope into the ownerData
of the Draw to save an object allocation/collection on
each draw pass.

Test: existing tests pass
Change-Id: I4229e21eb91596c4613b81544aea5917dca09538