Update patch set 11
Patch Set 11:

Change bypassed Presubmit.

Bypass Reason: other 
Additional Feedback: We're encountering a bunch of merge conflicts and test flakes as part of validating this change, and this change is essentially just flipping a flag and regenerating a bunch of files. One sample test failure is b/137137155 .

Because this change is essentially just flipping a flag and regenerating a bunch of files, if we do discover that there was anything problematic about this change, it should be straightforward enough to revert it.

Patch-set: 11
Tag: autogenerated:TreeHugger:Bypassed-Presubmit