Call onTrackInfoChanged for current media item only
onTrackInfoChanged is intended to be called when the current media
item is prepared but it may also be called when the next media item
is prepared because MediaPlayer2Impl may prepare the next item in
advance. This CL makes it as intended.

Additionally, since onTrackInfoChanged may not be called when the
current media item is changed (already prepared in the queue), it
makes widgets get new item's track info in onCurrentMediaItemChanged.

Bug: 137164160
Test: ./gradlew :media2:media2-player:connectedCheck \
                :media2:media2-widget:connectedCheck
      (on API 26/28 emulator, and 29 device)
Change-Id: I46922d820657f82d04b7b00761f5cd11c04ce219