Check platform accessibility delegate, not compat delegate for re-adding itemDelegates
To accomodate items being shared across different RecyclerViews ItemDelegates are removed on detach,
and readded on attach to a RecyclerView only when there isn't an app provided AccessibilityDelegate.
The app developer can still set their own accessibilityDelegate.
This check was broken, because although our guidance is that to a set a delegate on a recyclerView, you either use RecyclerViewAcccessibilityDelegate, or some of the VIewCompat functionaloty,
launcher was using the platform delegate (which we can't check for without new API in core). That's because calling ViewCompat.setAccessibilityDelegate(v, null), results in a default accessibility delegate set on the view.
So to make this potentially non uncommon case work, we directly set the platform accessibility delegate to null.

This has the consquence that using the ViewCOmpat accessibility APIs, while the view is detached (managed by the Recycler), will result in the item delegate not being re-added.
This kind of code would already break in other ways, so I am fine having it this way.

Change-Id: I2fdc32658b5375f6567c9f2dd5ce6195b343343f
Fix: 135773088
Test: Recycler Accessibility Tests, used logs to make sure item delegate is being re-added at the correct times.