Item delegate always added, refers to app developer added delegate
To accomodate items being shared across different RecyclerViews ItemDelegates are removed on detach,
and readded on attach to a RecyclerView only when there isn't an app provided AccessibilityDelegate.
The app developer can still set their own accessibilityDelegate.
This check was broken, because although our guidance is that to a set a delegate on a recyclerView, you either use RecyclerViewAcccessibilityDelegate, or some of the VIewCompat functionaloty,
launcher was using the platform delegate (which we can't check for without new API in core). That's because calling ViewCompat.setAccessibilityDelegate(v, null), results in a default accessibility delegate set on the view.
So to make this potentially non uncommon case work, we now have the item delegate keep track of the app provided delegate, and call into that, as well as the layout manager.
Ordered it so the app delegate can override teh item delegates default behavior in their own implementation.

Change-Id: I2fdc32658b5375f6567c9f2dd5ce6195b343343f
Fix: 135773088
Test: Recycler Accessibility Tests, modified them a nit, used logs to make sure item correct delegate is always called in the right order.