Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

If Doris likes the direction in general, I think we have enough to go on to move forward with this submission and refine as we go.  I think the direction is solid given that it is simply taking components from different pieces of functionality and bringing them together to achieve a more specific purpose.  Like I said, we may want to refine it... but that would clearly be done in a different CL.

Patch-set: 3
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1