Fix query writer not converting binding args to positional args.
Two issues are fixed in this CL:

1. When expanded projections are off, we still need to re-write the
a query with parameters, we failed to due this because the flag was too
broad.

2. Since bindings args got moved to the interpreter and the interpreter
is being used only read queries, then we stopped re-writing binding args
for write queries (prepared queries).

This CL fixes these two issues by correctly scoping the expandProjection
flag and adding the re-write to write queries.

The reason this was not caught in the tests is because DaoWriterTest
was changed to always interpret the query, which is not a reflection of
what the actual code is doing, I removed that and led the processor, as
it should, decide if the query needs to be re-written.

Bug: 137254857
Test: /gradlew room:room-compiler:test \
               room:integration-tests:room-testapp:cC

Change-Id: I3b354d12f2e916592244af0678993c141f46c77e