Various fixed regarding expanding projections.
1. When expanded projections are off, we still need to re-write the
a query with parameters, we failed to due this because the flag was too
broad.

2. Since bindings args got moved to the interpreter and the interpreter
is being used only read queries, then we stopped re-writing binding args
for write queries (prepared queries).

3. Database Views being transformed.

1 & 2 is fixed by correctly scoping the expandProjection flag and
adding the query transformation to write queries. The interpreted query
was also refactored to not be a public var.

3 is fixed by simply reverting the changes to expand projections in
views. Transforming the VIEW can be dangerous since it changes the
schema identity, which can lead to situations where we change our
transform logic and the schema identity.

Bug: 137254857
Test: /gradlew room:room-compiler:test \
               room:integration-tests:room-testapp:cC
Change-Id: I3b354d12f2e916592244af0678993c141f46c77e