Set mCurrentPrimaryItem to null it the current fragment is destroyed
In FragmentPagerAdapter and FragmentStatePagerAdapter if the current
Fragment is destroyed and mCurrentPrimaryItem is not set to null, it is
possible for the adapter to attempt to set the maxLifecycle and because
the Fragment has been removed an error is thrown.

When destroying items, we should check if the item being destroyed is
currently the primary item, and if it is, we should set
mCurrentPrimaryItem to null.

Test: tested in app
BUG: 137209870
Change-Id: I66e104160c9a15088dfc640e03a6afca021a5845
(cherry picked from commit 6f1170f62927702e491ac3aa7e55f68699d4c957)