Fix crash when a invalidation arrives late.
If an invalidation arrived for a part of the composition
was removed, it would cause a -1 to be entered into the
invalidations list of the composer causing the next
recompose to crash.

Also fixes an issue with animations that would crash if it
complete after the component being animated was removed from
the composition.

Test: New test, ./gradle test
Change-Id: Ib559b181069db1b24487b1709664fd48848d5136