Assign a sequential id to each TrackInfo
To avoid mis-selection of tracks for MediaPlayer, this CL sets id of
each TrackInfo with a sequence number instead of the index.

A side effect of this CL is that the list of tracks obtained from
MediaPlayer#getTrackInfo may become invalid within the same MediaItem
and users don't know the timing. It will be resolved soon when
SessionPlayer.PlayerCallback#onTrackInfoChanged becomes public.

Bug: 131873726
Test: ./gradlew :media2:media2-player:check \
                :media2:media2-player:connectedCheck (on API 26, 29)
Change-Id: I7cf816842881c458b7ff6ac45beaeb58e68f8982